refactor(pacbio-tag-store): created new store for tags in pacbio
